persistant & transient animations (+ doom fixes)
This adds an optional argument to start-animation and modifies the loop
behavior.  Whenever rotations is -1, only calling stop-animation will break it
out of this loop.

Show and hide were pulled out of the mode behavior and can be done independently
of the mode.

Adds `hide-when-not-animating' option that means to hide the parrot after
animation is finished.

See doom mentions in comments for doom.