Overhaul code structure, private
Added an option, parrot-animate-on-load.

Combined the frame list init code into one call.

Deconflated parrot-animation-state and parrot-hide-if-not-animating

Fixed a bug where I had regressed the handling of
parrot--old-cdr-mode-line-position

Added aliases where privates and moved variables will be missed by doom-modeline
or people using old customize settings.

Completely overhauled the order and naming to establish what is private to the
implementation as it is and what is likely being used by users.

Updated the `parrot--refresh` function to work with animations and initialize
properly, after the entire feature is loaded so that setters on customize and
elsewhere don't start calling things with half of package set up.